topic의 동적 생성은 되지만, springboot 서버를 다시 띄우지 않으면 컨슈머 그룹이 해당 토픽을 구독하지 못하는상태
