+++ src/main/java/com/hanghae/lemonairchat/kafka/KafkaTopicManager.java
@@ -4,8 +4,6 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.Properties;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutionException;
 
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.AdminClientConfig;
@@ -39,20 +37,17 @@ public class KafkaTopicManager {
 	public boolean isTopicCreated(String roomId) {
 		return topics.contains(roomId);
 	}
+
 	public Mono<Void> createTopic(String roomId, int partitions, short replicationFactor) {
 		return Mono.create(sink -> adminClient.listTopics().names().whenComplete((names, ex) -> {
-			log.info(String.valueOf(1));
 			if (ex != null) {
-				log.info("2");
 				sink.error(new RuntimeException(ex.getMessage()));
 				return;
 			}
 
 			if (names.contains(roomId)) {
-				log.info(String.valueOf(3));
 				sink.success();
 			} else {
-				log.info(String.valueOf(4));
 				NewTopic newTopic = new NewTopic(roomId, partitions, replicationFactor);
 				log.info(" {} 라는 새로운 토픽 생성 : ", roomId);
 				adminClient.createTopics(Collections.singleton(newTopic)).all().whenComplete((result, createEx) -> {
