+++ src/main/resources/application-common.yaml
@@ -14,3 +14,5 @@ server:
 logging:
   level:
     org.apache.kafka: info
+    root: info
+
